requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
django-ipware==2.1.0
gcloud==0.17.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jws==0.1.3
oauth2client==3.0.0
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-jwt==2.0.1
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.7.0
rsa==4.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.3.1
urllib3==1.25.9
whitenoise==5.0.1
zope.interface==5.0.1
